requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorlog==6.9.0
einops==0.8.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.4
hf-xet==1.1.10
huggingface-hub==0.35.0
hyperpyyaml==1.2.2
idna==3.10
jinja2==3.1.4
joblib==1.5.2
julius==0.2.7
lightning==2.5.5
lightning-utilities==0.15.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.3
numpy==2.3.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
omegaconf==2.3.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.0.0
primepy==1.3
propcache==0.3.2
protobuf==6.32.1
pyannote-audio==3.1.0
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pycparser==2.23
pygments==2.19.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
rich==14.1.0
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.12
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
sqlalchemy==2.0.43
sympy==1.13.3
tensorboardx==2.6.4
threadpoolctl==3.6.0
torch==2.8.0+cu129
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.8.0+cu129
torchmetrics==1.8.2
torchvision==0.23.0+cu129
tqdm==4.67.1
triton==3.4.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
